+++ MoPubSDK/Internal/Utility/MPGlobal.h
@@ -22,6 +22,8 @@ NSDictionary *MPDictionaryFromQueryString(NSString *query);
 NSString *MPSHA1Digest(NSString *string);
 BOOL MPViewIsVisible(UIView *view);
 BOOL MPViewIntersectsParentWindowWithPercent(UIView *view, CGFloat percentVisible);
+NSString *MPResourcePathForResource(NSString *resourceName);
+
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 /*
@@ -76,8 +78,6 @@ typedef NSUInteger MPInterstitialOrientationType;
 @interface UIDevice (MPAdditions)
 
 - (NSString *)hardwareDeviceName;
-- (BOOL)supportsOrientationMask:(UIInterfaceOrientationMask)orientationMask;
-- (BOOL)doesOrientation:(UIInterfaceOrientation)orientation matchOrientationMask:(UIInterfaceOrientationMask)orientationMask;
 
 @end
 
@@ -87,6 +87,8 @@ typedef NSUInteger MPInterstitialOrientationType;
 
 // Correct way to hide/show the status bar on pre-ios 7.
 - (void)mp_preIOS7setApplicationStatusBarHidden:(BOOL)hidden;
+- (BOOL)mp_supportsOrientationMask:(UIInterfaceOrientationMask)orientationMask;
+- (BOOL)mp_doesOrientation:(UIInterfaceOrientation)orientation matchOrientationMask:(UIInterfaceOrientationMask)orientationMask;
 
 @end