+++ Specs/MoPubSDKSpecs/Extensions/MRController+Specs.h
@@ -25,6 +25,8 @@
 @property (nonatomic, assign) CGSize currentAdSize;
 @property (nonatomic, assign) MRAdViewPlacementType placementType;
 @property (nonatomic, assign) UIInterfaceOrientation currentInterfaceOrientation;
+@property (nonatomic, copy) void (^forceOrientationAfterAnimationBlock)();
+@property (nonatomic, weak) MPMRAIDInterstitialViewController *interstitialViewController;
 
 - (void)checkViewability;
 - (void)orientationDidChange:(NSNotification *)notification;