+++ Specs/MoPubSDKSpecs/Fakes/FakeMPInstanceProvider.m
@@ -13,6 +13,7 @@
 #import "FakeMPTimer.h"
 #import "MRImageDownloader.h"
 #import "MRBundleManager.h"
+#import <FBAudienceNetwork/FBAudienceNetwork.h>
 
 @interface MPInstanceProvider (ThirdPartyAdditions)
 
@@ -26,6 +27,7 @@
 
 #pragma mark Facebook
 - (FBAdView *)buildFBAdViewWithPlacementID:(NSString *)placementID
+                                      size:(FBAdSize)size
                         rootViewController:(UIViewController *)controller
                                   delegate:(id<FBAdViewDelegate>)delegate;
 - (FBInterstitialAd *)buildFBInterstitialAdWithPlacementID:(NSString *)placementID
@@ -136,6 +138,30 @@
                      }];
 }
 
+#pragma mark - Rewarded Video
+- (MPRewardedVideoAdManager *)buildRewardedVideoAdManagerWithAdUnitID:(NSString *)adUnitID delegate:(id<MPRewardedVideoAdManagerDelegate>)delegate
+{
+    return [self returnFake:self.fakeMPRewardedVideoAdManager
+                     orCall:^{
+                         return [super buildRewardedVideoAdManagerWithAdUnitID:adUnitID delegate:delegate];
+                     }];
+}
+
+- (MPRewardedVideoAdapter *)buildRewardedVideoAdapterWithDelegate:(id<MPRewardedVideoAdapterDelegate>)delegate
+{
+    return [self returnFake:self.fakeMPRewardedVideoAdapter
+                     orCall:^{
+                         return [super buildRewardedVideoAdapterWithDelegate:delegate];
+                     }];
+}
+
+- (MPRewardedVideoCustomEvent *)buildRewardedVideoCustomEventFromCustomClass:(Class)aClass delegate:(id<MPRewardedVideoCustomEventDelegate>)delegate
+{
+    return [self returnFake:self.fakeMPRewardedVideoCustomEvent
+                     orCall:^{
+                         return [super buildRewardedVideoCustomEventFromCustomClass:aClass delegate:delegate];
+                     }];
+}
 #pragma mark - HTML Ads
 
 - (MPAdWebView *)buildMPAdWebViewWithFrame:(CGRect)frame delegate:(id<UIWebViewDelegate>)delegate
@@ -340,13 +366,13 @@
 
 #pragma mark - Facebook
 
-- (FBAdView *)buildFBAdViewWithPlacementID:(NSString *)placementID rootViewController:(UIViewController *)controller delegate:(id<FBAdViewDelegate>)delegate
+- (FBAdView *)buildFBAdViewWithPlacementID:(NSString *)placementID size:(FBAdSize)size rootViewController:(UIViewController *)controller delegate:(id<FBAdViewDelegate>)delegate
 {
     if (self.fakeFBAdView) {
         self.fakeFBAdView.delegate = delegate;
         return self.fakeFBAdView;
     } else {
-        return [super buildFBAdViewWithPlacementID:placementID rootViewController:controller delegate:delegate];
+        return [super buildFBAdViewWithPlacementID:placementID size:size rootViewController:controller delegate:delegate];
     }
 }