+++ Specs/MoPubSDKSpecs/Internal/MRAID/MRControllerSpec.mm
@@ -470,6 +470,12 @@ describe(@"MRController", ^{
 
             window = [[UIWindow alloc] initWithFrame:CGRectMake(0, 0, 320, 480)];
             UIViewController *rootViewController = [[UIViewController alloc] init];
+
+            // We're forcing this view controller into portrait with a hard coded frame size as resize tests rely heavily on the
+            // device's frame and we do not want outside orientation settings affecting the frame.
+            spy_on(rootViewController);
+            rootViewController stub_method(@selector(supportedInterfaceOrientations)).and_return(UIInterfaceOrientationMaskPortrait);
+
             [rootViewController.view addSubview:controller.mraidAdView];
             window.rootViewController = rootViewController;
             [window makeKeyWindow];
@@ -735,6 +741,13 @@ describe(@"MRController", ^{
             it(@"should not load anything", ^{
                 URL = [NSURL URLWithString:@"mopub://close"];
                 [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther] should equal(NO);
+            });
+
+            it(@"should notify the delegate when the url is failLoad", ^ {
+                URL = [NSURL URLWithString:@"mopub://failLoad"];
+                [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther] should equal(NO);
+
+                controllerDelegate should have_received(@selector(adDidFailToLoad:));
             });
         });
 
@@ -1176,6 +1189,171 @@ describe(@"MRController", ^{
                 [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther];
 
                 pictureManager should_not have_received(@selector(storePicture:));
+            });
+        });
+
+        context(@"when the command is 'setOrientationProperties", ^{
+            __block UIApplication *sharedApplication;
+            __block UIViewController *rootViewController;
+
+            beforeEach(^{
+                sharedApplication = [UIApplication sharedApplication];
+                spy_on(sharedApplication);
+
+                rootViewController = [[UIViewController alloc] init];
+                window.rootViewController = rootViewController;
+                [window makeKeyWindow];
+
+                URL = [NSURL URLWithString:@"mraid://setOrientationProperties?allowOrientationChange=false&forceOrientation=landscape"];
+            });
+
+            context(@"when our ad is an expanded banner", ^{
+                __block MRExpandModalViewController *expandViewController;
+
+                beforeEach(^{
+                    controller.placementType = MRAdViewPlacementTypeInline;
+                    controller.currentState = MRAdViewStateExpanded;
+                    controller.mraidBridge.shouldHandleRequests = YES;
+
+                    expandViewController = [[MRExpandModalViewController alloc] init];
+                    controller.expandModalViewController = expandViewController;
+
+                    [rootViewController presentViewController:expandViewController animated:NO completion:NULL];
+                });
+
+                context(@"when the app doesn't support the force orientation", ^{
+                    beforeEach(^{
+                        sharedApplication stub_method(@selector(mp_supportsOrientationMask:)).and_return(NO);
+
+                        [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther];
+                    });
+
+                    it(@"should not attempt to force the orientation", ^{
+                        controller should_not have_received(@selector(presentExpandModalViewControllerWithView:animated:completion:));
+                    });
+                });
+
+                context(@"when the app supports the force orientation", ^{
+                    beforeEach(^{
+                        sharedApplication stub_method(@selector(mp_supportsOrientationMask:)).and_return(YES);
+                    });
+
+                    it(@"should not attempt to force the orientation", ^{
+                        [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther];
+                        controller should have_received(@selector(presentExpandModalViewControllerWithView:animated:completion:));
+                    });
+
+                    it(@"should nil out the forceOrientationAfterAnimationBlock", ^{
+                        // Give the controller a random forceOrientationAfterAnimationBlock so we can determine if it gets set to nil when setOrientationProperties is called.
+                        controller.forceOrientationAfterAnimationBlock = ^{
+                            NSLog(@"hello");
+                        };
+
+                        [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther];
+                        controller.forceOrientationAfterAnimationBlock should be_nil;
+                    });
+
+                    context(@"when the ad is currently animating", ^{
+                        beforeEach(^{
+                            controller.mraidBridge.shouldHandleRequests = NO;
+                            [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther];
+                        });
+
+                        it(@"should create a block to execute the force orientation", ^{
+                            controller.forceOrientationAfterAnimationBlock should_not be_nil;
+                        });
+
+                        it(@"should execute the block after the animation completes", ^{
+                            [(id<CedarDouble>)controller reset_sent_messages];
+                            controller should_not have_received(@selector(bridge:handleNativeCommandSetOrientationPropertiesWithForceOrientationMask:));
+                            [controller enableRequestHandling];
+                            [controller handleMRAIDInterstitialDidPresentWithViewController:nil];
+                            controller should have_received(@selector(bridge:handleNativeCommandSetOrientationPropertiesWithForceOrientationMask:));
+                            controller.forceOrientationAfterAnimationBlock should be_nil;
+                        });
+                    });
+
+                });
+            });
+
+            context(@"when our ad type is interstitial", ^{
+                __block UIViewController<CedarDouble> *presentingViewController;
+                __block MPMRAIDInterstitialViewController *interstitialViewController;
+
+                beforeEach(^{
+                    presentingViewController = nice_fake_for([UIViewController class]);
+                    interstitialViewController = [[MPMRAIDInterstitialViewController alloc] init];
+                    spy_on(interstitialViewController);
+                    interstitialViewController stub_method(@selector(presentingViewController)).and_return(presentingViewController);
+
+                    controller.interstitialViewController = interstitialViewController;
+                    controller.placementType = MRAdViewPlacementTypeInterstitial;
+
+                    [rootViewController presentViewController:interstitialViewController animated:NO completion:NULL];
+                });
+
+                context(@"when the app doesn't support the force orientation", ^{
+                    beforeEach(^{
+                        sharedApplication stub_method(@selector(mp_supportsOrientationMask:)).and_return(NO);
+
+                        controller.mraidBridge.shouldHandleRequests = YES;
+
+                        [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther];
+                    });
+
+                    it(@"should not attempt to force the orientation", ^{
+                        // We dismiss and then present the same view controller to force orientation. So we make sure we don't observe
+                        // those methods after forcing the orientation.
+                        controller.interstitialViewController should_not have_received(@selector(dismissViewControllerAnimated:completion:));
+                        presentingViewController should_not have_received(@selector(presentViewController:animated:completion:));
+                    });
+                });
+
+                context(@"when the app supports the force orientation", ^{
+                    beforeEach(^{
+                        sharedApplication stub_method(@selector(mp_supportsOrientationMask:)).and_return(YES);
+
+                        controller.mraidBridge.shouldHandleRequests = YES;
+                    });
+
+                    it(@"should attempt to force the orientation", ^{
+                        [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther];
+
+                        // We dismiss and then present the same view controller to force orientation. We make sure we observe these methods to verify that we have forced orientation.
+                        controller.interstitialViewController should have_received(@selector(dismissViewControllerAnimated:completion:));
+                        presentingViewController should have_received(@selector(presentViewController:animated:completion:));
+                    });
+
+                    it(@"should nil out the forceOrientationAfterAnimationBlock", ^{
+                        // Give the controller a random forceOrientationAfterAnimationBlock so we can determine if it gets set to nil when setOrientationProperties is called.
+                        controller.forceOrientationAfterAnimationBlock = ^{
+                            NSLog(@"hello");
+                        };
+
+                        [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther];
+                        controller.forceOrientationAfterAnimationBlock should be_nil;
+                    });
+
+                    context(@"when the ad is currently animating", ^{
+                        beforeEach(^{
+                            controller.mraidBridge.shouldHandleRequests = NO;
+                            [fakeMRBridge webView:nil shouldStartLoadWithRequest:[NSURLRequest requestWithURL:URL] navigationType:UIWebViewNavigationTypeOther];
+                        });
+
+                        it(@"should create a block to execute the force orientation", ^{
+                            controller.forceOrientationAfterAnimationBlock should_not be_nil;
+                        });
+
+                        it(@"should execute the block after the animation completes", ^{
+                            [(id<CedarDouble>)controller reset_sent_messages];
+                            controller should_not have_received(@selector(bridge:handleNativeCommandSetOrientationPropertiesWithForceOrientationMask:));
+                            [controller enableRequestHandling];
+                            [controller handleMRAIDInterstitialDidPresentWithViewController:nil];
+                            controller should have_received(@selector(bridge:handleNativeCommandSetOrientationPropertiesWithForceOrientationMask:));
+                            controller.forceOrientationAfterAnimationBlock should be_nil;
+                        });
+                    });
+                });
             });
         });
     });